pyproject.toml
[tool.poetry]
name = "appstore"
version = "0.1.0"
description = "Downloading review data via App Store Connect."
authors = ["Hamed Bastan-Hagh <hamed.bastan-hagh@hmrc.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
Authlib = "^1.3.1"
requests = "^2.32.3"
tqdm = "^4.66.5"
python-dotenv = "^1.0.1"
poetry-plugin-export = "^1.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"
pytest = "^8.3.2"
poetry-plugin-export = "^1.8.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.lint]
# The default set of rules is "E4", "E7", "E9", "F"
select = ["E4", "E7", "E9", "F", "I"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default in ruff, but planned
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
